Move logic into module and wire everything up
