+++ crates/autopilot/src/domain/eth/mod.rs
@@ -252,7 +252,21 @@ pub struct Asset {
 }
 
 /// An amount of native Ether tokens denominated in wei.
-#[derive(Clone, Copy, Debug, Eq, Ord, PartialEq, PartialOrd, From, Into, Display, Default, derive_more::AddAssign, derive_more::Add)]
+#[derive(
+    Clone,
+    Copy,
+    Debug,
+    Eq,
+    Ord,
+    PartialEq,
+    PartialOrd,
+    From,
+    Into,
+    Display,
+    Default,
+    derive_more::AddAssign,
+    derive_more::Add,
+)]
 pub struct Ether(pub U256);
 
 impl num::Zero for Ether {
