+++ crates/autopilot/src/run_loop.rs
@@ -10,8 +10,8 @@ use {
                 Solution,
                 SolutionError,
                 SolverParticipationGuard,
-                TradedOrder,
                 Unranked,
+                winner_selection::{self, Arbitrator},
             },
             eth::{self, TxId},
             settlement::{ExecutionEnded, ExecutionStarted},
@@ -27,7 +27,6 @@ use {
     ::observe::metrics,
     anyhow::Result,
     database::order_events::OrderEventLabel,
-    ethcontract::{H160, U256},
     ethrpc::block_stream::BlockInfo,
     futures::{FutureExt, TryFutureExt},
     itertools::Itertools,
@@ -43,6 +42,7 @@ use {
     shared::token_list::AutoUpdatingTokenList,
     std::{
         collections::{HashMap, HashSet},
+        num::NonZeroUsize,
         sync::Arc,
         time::{Duration, Instant},
     },
@@ -58,8 +58,8 @@ pub struct Config {
     /// allowed to start before it has to re-synchronize to the blockchain
     /// by waiting for the next block to appear.
     pub max_run_loop_delay: Duration,
-    pub max_winners_per_auction: usize,
-    pub max_solutions_per_solver: usize,
+    pub max_winners_per_auction: NonZeroUsize,
+    pub max_solutions_per_solver: NonZeroUsize,
 }
 
 pub struct RunLoop {
@@ -76,6 +76,7 @@ pub struct RunLoop {
     /// the most recent data available.
     maintenance: Arc<Maintenance>,
     competition_updates_sender: tokio::sync::mpsc::UnboundedSender<()>,
+    winner_selection: Box<dyn Arbitrator>,
 }
 
 impl RunLoop {
@@ -93,6 +94,14 @@ impl RunLoop {
         competition_updates_sender: tokio::sync::mpsc::UnboundedSender<()>,
     ) -> Self {
         Self {
+            winner_selection: match config.max_winners_per_auction.get() {
+                0 => unreachable!(),
+                1 => Box::new(winner_selection::max_surplus::Config),
+                n => Box::new(winner_selection::combinatorial::Config {
+                    max_winners: n,
+                    weth: eth.contracts().wrapped_native_token(),
+                }),
+            },
             config,
             eth,
             persistence,
@@ -240,11 +249,14 @@ impl RunLoop {
             .store_order_events(auction.orders.iter().map(|o| o.uid), OrderEventLabel::Ready);
 
         // Collect valid solutions from all drivers
-        let solutions = self.competition(&auction).await;
+        let solutions = self.fetch_solutions(&auction).await;
         observe::solutions(&solutions);
         if solutions.is_empty() {
             return;
         }
+
+        let solutions = self.winner_selection.filter_solutions(solutions, &auction);
+        let solutions = self.winner_selection.mark_winners(solutions);
 
         let competition_simulation_block = self.eth.current_block().borrow().number;
         let block_deadline = competition_simulation_block + self.config.submission_deadline;
@@ -550,374 +562,12 @@ impl RunLoop {
             let driver = participant.driver().name.clone();
             let count = counter.entry(driver).or_insert(0);
             *count += 1;
-            *count <= self.config.max_solutions_per_solver
+            *count <= self.config.max_solutions_per_solver.get()
         });
 
         // Shuffle so that sorting randomly splits ties.
         solutions.shuffle(&mut rand::thread_rng());
         solutions
-    }
-
-    /// Filter out solutions that are not fair
-    async fn filter_solutions_old(
-        &self,
-        solutions: Vec<competition::Participant<Unranked>>,
-        auction: &domain::Auction,
-    ) -> Vec<competition::Participant<Unranked>> {
-        solutions
-            .iter()
-            .enumerate()
-            .filter_map(|(index, participant)| {
-                if Self::is_solution_fair(participant, &solutions[index..], auction) {
-                    Some(participant.clone())
-                } else {
-                    tracing::warn!(
-                        invalidated = participant.driver().name,
-                        "fairness check invalidated of solution"
-                    );
-                    None
-                }
-            })
-            .collect()
-    }
-
-    /// Drops all solutions where the score for any directed token pair is less than
-    /// what the baseline solution provides for that token pair.
-    /// If no baseline solution exists for a token pair a score of 0 is assumed.
-    fn filter_solutions_new(
-        &self,
-        solutions: Vec<competition::Participant<Unranked>>,
-    ) -> Vec<competition::Participant<Unranked>> {
-        // drops all solutions that have at least 1 trade execution which can't beat the best
-        // baseline trade execution for that directed token pair.
-        let mut best_single_solutions = HashMap::<(H160, H160), (usize, Score)>::default();
-        for solution in solutions {
-
-        }
-        todo!()
-    }
-
-    /// Let's call a solution that only trades 1 directed token pair a baseline solution.
-    /// Returns the best baseline solution (highest score) for each token pair if one
-    /// exists.
-    fn compute_baseline_solutions(solutions: &[competition::Participant<Unranked>]) -> HashMap<(eth::TokenAddress, eth::TokenAddress), &competition::Participant<Unranked>> {
-        let mut baseline_solutions = HashMap::default();
-        for solution in solutions {
-            let aggregate_scores = solution.solution().aggregate_scores(());
-            if aggregate_scores.len() != 1 {
-                // base solutions must contain exactly 1 directed token pair
-                continue;
-            }
-            let (token_pair, score) = aggregate_scores.into_iter().next().unwrap();
-            let current_best_score = baseline_solutions.entry(token_pair).or_default();
-            if score > *current_best_score {
-                *current_best_score = score;
-            }
-        }
-        baseline_solutions
-    }
-
-    fn mark_winners_old(
-        &self,
-        mut solutions: Vec<competition::Participant<Unranked>>,
-    ) -> Vec<competition::Participant> {
-        // The current system theoretically already supports multiple winners. However,
-        // it was never activated because the rewards mechanism was never
-        // decided. To make the migration easier we revert back to only allowing
-        // 1 winner. And that is simply the solution with the highest total
-        // score.
-        solutions.sort_unstable_by_key(|participant| {
-            std::cmp::Reverse(participant.solution().score().get().0)
-        });
-        solutions
-            .into_iter()
-            .enumerate()
-            .map(|(index, solution)| solution.rank(index == 0))
-            .collect()
-    }
-
-    fn mark_winners_new(
-        &self,
-        mut solutions: Vec<competition::Participant<Unranked>>,
-    ) -> Vec<competition::Participant> {
-        solutions.sort_unstable_by_key(|participant| {
-            std::cmp::Reverse(participant.solution().score().get().0)
-        });
-
-        // Winners are selected one by one, starting from the best solution,
-        // until `max_winners_per_auction` are selected. A solution can only
-        // win if none of the (sell_token, buy_token) pairs of the executed
-        // orders have been covered by any previously selected winning solution.
-        // In other words this enforces a uniform directional clearing price.
-        let wrapped_native_token = self.eth.contracts().wrapped_native_token();
-        let mut already_swapped_tokens_pairs = HashSet::new();
-        let mut winners = 0;
-        solutions
-            .into_iter()
-            .map(|participant| {
-                let swapped_token_pairs = participant
-                    .solution()
-                    .orders()
-                    .iter()
-                    .map(|(_, order)| {
-                        (
-                            order.sell.token.as_erc20(wrapped_native_token),
-                            order.buy.token.as_erc20(wrapped_native_token),
-                        )
-                    })
-                    .collect::<HashSet<_>>();
-
-                let is_winner = swapped_token_pairs.is_disjoint(&already_swapped_tokens_pairs)
-                    && winners < self.config.max_winners_per_auction;
-
-                already_swapped_tokens_pairs.extend(swapped_token_pairs);
-                winners += usize::from(is_winner);
-
-                participant.rank(is_winner)
-            })
-            .collect()
-    }
-
-    fn compute_reference_scores_new(
-        &self,
-        solutions: &[competition::Participant],
-    ) -> HashMap<eth::Address, competition::Score> {
-        let mut reference_scores = HashMap::default();
-
-        for solution in solutions {
-            let driver = solution.driver().submission_address;
-            if reference_scores.contains_key(&driver) {
-                // we already computed the reference score
-                continue;
-            }
-
-            let solutions_without_solver = solutions
-                .iter()
-                .filter(|s| s.driver().submission_address != driver)
-                .cloned()
-                .map(|solution| solution.unrank())
-                .collect();
-            let ranked = self.mark_winners_new(solutions_without_solver);
-            let score = ranked
-                .iter()
-                .filter(|s| s.is_winner())
-                .fold(U256::zero(), |acc, s| acc + s.solution().score().0);
-            let score = competition::Score::try_new(eth::Ether(score)).unwrap_or_default();
-            reference_scores.insert(driver, score);
-        }
-
-        reference_scores
-    }
-
-    async fn compute_reference_scores_old(
-        &self,
-        solutions: &[competition::Participant],
-    ) -> HashMap<eth::Address, competition::Score> {
-        // this will hold at most 1 score but the interface needs to support multiple
-        // scores to fit the interface
-        let mut reference_scores = HashMap::default();
-        if let Some(winner) = solutions.get(0) {
-            let runner_up = solutions
-                .get(1)
-                .map(|s| s.solution().score())
-                .unwrap_or_default();
-            reference_scores.insert(winner.driver().submission_address, runner_up);
-        }
-        reference_scores
-    }
-
-    /// Runs the solver competition, making all configured drivers participate.
-    /// Returns all fair solutions sorted by their score (best to worst).
-    async fn competition(&self, auction: &domain::Auction) -> Vec<competition::Participant> {
-        let request = solve::Request::new(
-            auction,
-            &self.trusted_tokens.all(),
-            self.config.solve_deadline,
-        );
-        let request = &request;
-
-        let mut solutions = futures::future::join_all(
-            self.drivers
-                .iter()
-                .map(|driver| self.solve(driver.clone(), request)),
-        )
-        .await
-        .into_iter()
-        .flatten()
-        .collect::<Vec<_>>();
-
-        // Shuffle so that sorting randomly splits ties.
-        solutions.shuffle(&mut rand::thread_rng());
-
-        solutions.sort_unstable_by_key(|participant| {
-            std::cmp::Reverse(participant.solution().score().get().0)
-        });
-
-        // Filter out solutions that don't come from their corresponding submission
-        // address
-        let mut solutions = solutions
-            .into_iter()
-            .filter(|participant| {
-                let submission_address = participant.driver().submission_address;
-                let is_solution_from_driver = participant.solution().solver() == submission_address;
-                if !is_solution_from_driver {
-                    tracing::warn!(
-                        driver = participant.driver().name,
-                        ?submission_address,
-                        "the solution received is not from the driver submission address"
-                    );
-                }
-                is_solution_from_driver
-            })
-            .collect::<Vec<_>>();
-
-        // Limit the number of accepted solutions per solver. Do not alter the ordering
-        // of solutions
-        let mut counter = HashMap::new();
-        solutions.retain(|participant| {
-            let driver = participant.driver().name.clone();
-            let count = counter.entry(driver).or_insert(0);
-            *count += 1;
-            *count <= self.config.max_solutions_per_solver
-        });
-
-        // this will be done differently with CA
-        // Filter out solutions that are not fair
-        let solutions = solutions
-            .iter()
-            .enumerate()
-            .filter_map(|(index, participant)| {
-                if Self::is_solution_fair(participant, &solutions[index..], auction) {
-                    Some(participant)
-                } else {
-                    tracing::warn!(
-                        invalidated = participant.driver().name,
-                        "fairness check invalidated of solution"
-                    );
-                    None
-                }
-            });
-
-        // this will be done differently with CA
-        // Winners are selected one by one, starting from the best solution,
-        // until `max_winners_per_auction` are selected. The solution is a winner
-        // if it swaps tokens that are not yet swapped by any previously processed
-        // solution.
-        let wrapped_native_token = self.eth.contracts().wrapped_native_token();
-        let mut already_swapped_tokens = HashSet::new();
-        let mut winners = 0;
-        let solutions = solutions
-            .cloned()
-            .map(|participant| {
-                let swapped_tokens = participant
-                    .solution()
-                    .orders()
-                    .iter()
-                    .flat_map(|(_, order)| {
-                        [
-                            order.sell.token.as_erc20(wrapped_native_token),
-                            order.buy.token.as_erc20(wrapped_native_token),
-                        ]
-                    })
-                    .collect::<HashSet<_>>();
-
-                let is_winner = swapped_tokens.is_disjoint(&already_swapped_tokens)
-                    && winners < self.config.max_winners_per_auction;
-
-                already_swapped_tokens.extend(swapped_tokens);
-                winners += usize::from(is_winner);
-
-                participant.rank(is_winner)
-            })
-            .collect();
-
-        solutions
-    }
-
-    /// Returns true if solution is fair to other solutions
-    fn is_solution_fair(
-        solution: &competition::Participant<Unranked>,
-        others: &[competition::Participant<Unranked>],
-        auction: &domain::Auction,
-    ) -> bool {
-        let Some(fairness_threshold) = solution.driver().fairness_threshold else {
-            return true;
-        };
-
-        // Returns the surplus difference in the buy token if `left`
-        // is better for the trader than `right`, or 0 otherwise.
-        // This takes differently partial fills into account.
-        let improvement_in_buy = |left: &TradedOrder, right: &TradedOrder| {
-            // If `left.sell / left.buy < right.sell / right.buy`, left is "better" as the
-            // trader either sells less or gets more. This can be reformulated as
-            // `right.sell * left.buy > left.sell * right.buy`.
-            let right_sell_left_buy = right.executed_sell.0.full_mul(left.executed_buy.0);
-            let left_sell_right_buy = left.executed_sell.0.full_mul(right.executed_buy.0);
-            let improvement = right_sell_left_buy
-                .checked_sub(left_sell_right_buy)
-                .unwrap_or_default();
-
-            // The difference divided by the original sell amount is the improvement in buy
-            // token. Casting to U256 is safe because the difference is smaller than the
-            // original product, which if re-divided by right.sell must fit in U256.
-            improvement
-                .checked_div(right.executed_sell.0.into())
-                .map(|v| U256::try_from(v).expect("improvement in buy fits in U256"))
-                .unwrap_or_default()
-        };
-
-        // Record best execution per order
-        let mut best_executions = HashMap::new();
-        for other in others {
-            for (uid, execution) in other.solution().orders() {
-                best_executions
-                    .entry(uid)
-                    .and_modify(|best_execution| {
-                        if !improvement_in_buy(execution, best_execution).is_zero() {
-                            *best_execution = *execution;
-                        }
-                    })
-                    .or_insert(*execution);
-            }
-        }
-
-        // Check if the solution contains an order whose execution in the
-        // solution is more than `fairness_threshold` worse than the
-        // order's best execution across all solutions
-        let unfair = solution
-            .solution()
-            .orders()
-            .iter()
-            .any(|(uid, current_execution)| {
-                let best_execution = best_executions.get(uid).expect("by construction above");
-                let improvement = improvement_in_buy(best_execution, current_execution);
-                if improvement.is_zero() {
-                    return false;
-                };
-                tracing::debug!(
-                    ?uid,
-                    ?improvement,
-                    ?best_execution,
-                    ?current_execution,
-                    "fairness check"
-                );
-                // Improvement is denominated in buy token, use buy price to normalize the
-                // difference into eth
-                let Some(order) = auction.orders.iter().find(|order| order.uid == *uid) else {
-                    // This can happen for jit orders
-                    tracing::debug!(?uid, "cannot ensure fairness, order not found in auction");
-                    return false;
-                };
-                let Some(buy_price) = auction.prices.get(&order.buy.token) else {
-                    tracing::warn!(
-                        ?order,
-                        "cannot ensure fairness, buy price not found in auction"
-                    );
-                    return false;
-                };
-                buy_price.in_eth(improvement.into()) > fairness_threshold
-            });
-        !unfair
     }
 
     /// Sends a `/solve` request to the driver and manages all error cases and
@@ -1367,7 +1017,10 @@ impl Metrics {
 }
 
 pub mod observe {
-    use {crate::domain, std::collections::HashSet};
+    use {
+        crate::domain::{self, competition::Unranked},
+        std::collections::HashSet,
+    };
 
     pub fn log_auction_delta(previous: &Option<domain::Auction>, current: &domain::Auction) {
         let previous_uids = match previous {
@@ -1397,7 +1050,7 @@ pub mod observe {
         );
     }
 
-    pub fn solutions(solutions: &[domain::competition::Participant]) {
+    pub fn solutions(solutions: &[domain::competition::Participant<Unranked>]) {
         if solutions.is_empty() {
             tracing::info!("no solutions for auction");
         }
